SAK-2807 Remove bogus warning when a Gradebook-less site is deleted from Sakai
git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai/gradebook@3884 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ gradebook/tool/src/java/org/sakaiproject/tool/gradebook/facades/sakai2impl/EntityProducerSakai2.java
@@ -72,7 +72,7 @@ public class EntityProducerSakai2 implements EntityProducer {
 				if (log.isInfoEnabled()) log.info("Gradebook being added to site " + gradebookUid);
 				gradebookService.addGradebook(gradebookUid, gradebookUid);
 			}
-		} else if (change == EntityProducer.ChangeType.REMOVE) {
+		} else if ((change == EntityProducer.ChangeType.REMOVE) && gradebookExists) {
 			try {
 				gradebookService.deleteGradebook(gradebookUid);
 			} catch (GradebookNotFoundException e) {
